+++ src/review/reivew.interface.ts
@@ -1,0 +1,4 @@
+export interface FindReviewOptionsProps {
+  boothName: string;
+  nickname: string;
+}
